Bug fix: Inline object instantiations within methods for scope safety
When instantiating an object inside of a method, rather than calling the class's 'new' function externally, we can inline the contents of that 'new' function in order to ensure that the created object exists purely within a local scope (declared 'local' in the generated Bash)
This fixes a bug with recursive method calls in Bash++, where an instantiated object which was not declared as a pointer (and therefore did not receive a runtime-randomized allocation from 'new', but received a deterministic allocation determined at compile-time) would be overwritten by subsequent calls to the same method which would use the same compile-determined address
It may be better in the long-run to ensure that all created objects receive runtime-randomized allocations from 'new' (see #6)

This change has also introduced a fair bit of code duplication which will soon need to be cleaned up. The portion of bpp_program::add_class which is responsible for generating the class's 'new' function has been duplicated in bpp_codegen's 'inline_new'. bpp_method now also overrides bpp_code_entity::add_object with identical code, save the difference that we call the new 'inline_new' function. This will be cleaned up soon.