Bug fix: Allow accessing the class from within the class
Previously, a hacky fix had been implemented to allow specifically pointers to the same class type as data members of a class
However, it was still impossible to reference a class from within itself in other contexts -- eg, from within a method

The most reasonable to this would have been to modify the get_class(std::string name) function of bpp_entity to, once it discovered that the given name did not exist in its classes map, to check if the name matched that of the containing_class weak_ptr, and if so, return containing_class.lock()
However, it was found that the weak_ptr in question would become invalidated from time to time. After a lot of debugging, I'm convinced that this is a bug in the C++ standard library implementation -- the spec says that it should work.
So, as a consequence of that, we have another semi-hacky solution: bpp_program::prepare_class(). This function adds the class to the program's classes map without generating its code. When we enter a class definition now, we call program->prepare_class() immediately, and call program->add_class() when we exit the class definition (only at that point do we have enough information to generate its code). But, adding it to the program's classes map before we finish parsing it allows constructs inside of the class to reference the class itself. It is however a shame that the smart pointers do not work as the spec describes, and this should ideally be revisited later.