Bug fix: Maintain one global list of includes
This fixes a particular bug, where, for example:
  - File A includes File B using include_once
  - File B includes File C using include_once
  - We return from parsing File B
  - File A includes File C using include_once

With each file having its own include list, 'C' would've been added to B's includes but not A's. So A wouldn't have known not to try including it again.
Now we have a single global include list, and each parse holds a pointer to it