Bug fix: Properly scan the entity stack for the latest code entity
In some cases, when trying to get the relevant code entity from within the parser, we checked if the top of the entity stack was a code entity, and if not, set the relevant code entity to the bpp::program.

This led to bugs in specific cases such as the following:
@Object* obj=@new Object
@Object* obj2=@obj

If those two lines were in the global program scope, there was no problem. However, if they existed in some kind of code entity lower than the program scope, then the second one would fail. Why?
Because the object reference to @obj is inside a value assignment (=...) and a pointer declaration. Neither of those two are code entities.
So when parsing the object reference, we mistakenly set the relevant code entity to the global program, and then searched *that* code entity for an object called 'obj' -- and of course, didn't find it. It wasn't declared in the global program scope -- it was declared where *we* are.

Now we have a function under BashppListener::latest_code_entity() which traverses the entity stack until it finds the most recent code entity and returns it.