Bug fix in entity resolution: Declare all temporary variables local if possible
If we're inside a class, a supershell, or bash function, it's possible for us to declare all temporary variables 'local', and we absolutely should
This fixes a bug with nested object references. For example, @this.member=@(echo @this.member and then some)
If the temporary variables necessary to access this.member are not declared local, then:
 - The pre-code necessary to access this.member precedes that assignment statement
 - Then, we enter the supershell in the rvalue of the assignment statement, which also contains a reference to this.member
 - So, once again we generate the pre-code necessary to access it, we access it,
 - AND THEN we execute the post-code necesary to clear the memory
 - That memory having been cleared, we return from the supershell context to the broader assignment statement
And we find that we no longer know where to put the rvalue. The variables which told us how to access this.member have been cleared from inside the supershell. Long story short, the assignment statement fails

If we declare all temporary variables local however, then the pre- and post-code within the supershell is entirely isolated from the pre- and post-code of the surrounding assignment statement.

This change fixes that bug and in general isolates temporary runtime variables to their relevant scope