Stateful considerations: Determine acceptable entity types in context, before returning from resolve_reference
This is another step in the effort to reduce the 4x code duplication across the 2^2 types of entity reference handlers (object/self, and lvalue/rvalue) by abstracting the core functionality to generalized functions in the compiler.

The first step along that road was to abstract the logic for figuring out *which* entity was being referred to by a sequence of identifiers. This became the resolve_reference function in bpp_codegen.

This step expands that function slightly (and maybe inappropriately) by moving the responsibility for *contextualization* to that same function. I.e., where previously we may have parsed a statement such as:
   echo @obj
By (1) passing 'obj' to resolve_reference, which would return a pointer to the nonprimitive object 'obj',
Then (2) AFTER receiving that pointer, determining from context that a primitive is expected,
And finally (3) expanding the reference to an implicit call to '@obj.toPrimitive'
  (This method, incidentally, meant that the expansion/context checking had to be done SEPARATELY for EACH of the 4 entity reference handler functions)
Now, instead, the context considerations are taken care of **inside** of resolve_reference.

The listener class keeps track of stateful variables 'can_take_object' and 'can_take_primitive' as we traverse the parse tree, which can then be quickly checked by resolve_reference to determine whether what's been given matches what's expected, and if not, what to do about it.

The next step along this road is to try and generalize the remaining parts (the proper 'code-writing' parts) of the entity reference handlers to a single abstracted function. Then finally some cleanup is due to make sure everything's sane. Some of the architecture is confusing as it is already and needs to be re-thought.