Move context expectations storage into a stack
The previous config (single boolean values) is bug-prone. Say we enter context A which sets expectation set 1. Then, inside ctx A we enter ctx B which sets expectation set 2.
In the previous model: exiting ctx B would immediately reset the expectation set to the default, losing ctx A's special expectation set. This could lead to bugs.
In the current model: exiting ctx B pops its expectation set off the stack, restoring the expectation set from ctx A as we return to ctx A. Later, when we exit ctx A, its expectations are also popped off the stack