Update resolve_entity call in language server to match new context-aware syntax
Just say that both primitives and nonprimitives are acceptable s.t. we get back whatever the first match is without adulteration. Obviously if we were to say, for example, that direct nonprimitives were unacceptable than a Hover Request over an object would return the .toPrimitive method, which is obviously not what we want